EIEIOParameters class in its own file
